Implement middleware orchestration logic
This is essentially a first draft over the logic which determines when
middleware should run, and with what arguments.

Also included is a sample middleware, which prepends `[Mineshark]` to
the front of server status messages. When a connection is successfully
established, which is not guaranteed atm, this middleware intercepts
packets from the server and changes the server description text.