Remove encryption logic
I've just discovered that notchian servers have some protection against
man-in-the-middle attacks when online mode is enabled. This means that
for Mineshark to work, the target server needs to be set to offline
mode.

This has the unfortunate consequence of also disabling encryption. So
I've stripped out anything encryption related since it is no longer
needed.